+++ embed-pdf-viewer.php
@@ -14,7 +14,7 @@
  * Description:       Embed a PDF from the Media Library or elsewhere via oEmbed or as a block into an `object` tag or Google Doc Viewer as fallback.
  * Author:            Andy Fragen
  * Author URI:        https://github.com/afragen
- * Version:           2.0.5
+ * Version:           2.1.0
  * License:           GPLv2+
  * Domain Path:       /languages
  * Text Domain:       embed-pdf-viewer
@@ -45,6 +45,7 @@ add_action(
 	function () {
 		load_plugin_textdomain( 'embed_pdf_viewer' );
 
+		// phpcs:ignore WordPress.WP.EnqueuedResourceParameters.NoExplicitVersion
 		wp_enqueue_style(
 			'embed-pdf-viewer',
 			plugins_url( 'css/embed-pdf-viewer.css', __FILE__ ),
@@ -90,9 +91,10 @@ class Embed_PDF_Viewer {
 			return;
 		}
 
+		// phpcs:ignore WordPress.WP.EnqueuedResourceParameters.NoExplicitVersion
 		wp_register_script(
 			'embed-pdf-viewer',
-			plugins_url( 'blocks/index.build.js', __FILE__ ),
+			plugins_url( 'blocks/build/index.js', __FILE__ ),
 			[ 'wp-blocks', 'wp-element' ],
 			false,
 			true
@@ -164,9 +166,10 @@ class Embed_PDF_Viewer {
 		}
 
 		$default = [
-			'height' => 500,
-			'width'  => 800,
-			'title'  => $post->post_title,
+			'height'      => 500,
+			'width'       => 800,
+			'title'       => $post->post_title,
+			'description' => $post->post_content,
 		];
 
 		/*
@@ -189,31 +192,27 @@ class Embed_PDF_Viewer {
 		$atts = apply_filters( 'embed_pdf_viewer_pdf_attributes', $atts );
 
 		// Fix title or create from filename.
-		$atts['title'] = empty( $atts['title'] )
+		$atts['title']       = empty( $atts['title'] )
 			? ucwords( preg_replace( '/(-|_)/', ' ', $post->post_name ) )
 			: ucwords( preg_replace( '/(-|_)/', ' ', $atts['title'] ) );
+		$atts['description'] = empty( $atts['description'] ) ? $atts['title'] : $atts['description'];
 
 		$iframe_fallback  = '<iframe class="embed-pdf-viewer" src="https://docs.google.com/viewer?url=' . rawurlencode( $post->guid );
 		$iframe_fallback .= '&amp;embedded=true" frameborder="0" ';
 		$iframe_fallback .= 'style="height:' . $atts['height'] . 'px;width:' . $atts['width'] . 'px;" ';
-		$iframe_fallback .= 'title="' . $atts['title'] . '"></iframe>' . "\n";
+		$iframe_fallback .= 'title="' . $atts['description'] . '"></iframe>' . "\n";
 
 		$object  = '<object class="embed-pdf-viewer" data="' . $post->guid;
 		$object .= '#scrollbar=1&toolbar=1"';
 		$object .= 'type="application/pdf" ';
-		$object .= 'height=' . $atts['height'] . ' width=' . $atts['width'] . ' > ';
+		$object .= 'height=' . $atts['height'] . ' width=' . $atts['width'] . ' ';
+		$object .= 'title="' . $atts['description'] . '"> ';
 		$object .= '</object>';
 
-		$style = '<style>
-			@media only screen and (min-device-width: 1024px) {
-				iframe.embed-pdf-viewer { display:none; }
-				object.embed-pdf-viewer { display:block; }
-			}
-		</style>';
-
-		$embed  = $object;
-		$embed .= $style . $iframe_fallback;
-		$embed .= '<p><a href="' . $post->guid . '">' . $atts['title'] . '</a></p>';
+		$embed  = '<figure>';
+		$embed .= $object . $iframe_fallback;
+		$embed .= '<p><a href="' . $post->guid . '" title="' . $atts['description'] . '">' . $atts['title'] . '</a></p>';
+		$embed .= '</figure>';
 
 		return $embed;
 	}
@@ -229,6 +228,7 @@ class Embed_PDF_Viewer {
 	 */
 	private function get_attachment_id_by_url( $url ) {
 		global $wpdb;
+		// phpcs:ignore WordPress.DB
 		$attachment = $wpdb->get_col( $wpdb->prepare( "SELECT ID FROM $wpdb->posts WHERE guid='%s';", $url ) );
 
 		if ( empty( $attachment ) ) {