+++ CHANGES.md
@@ -1,5 +1,10 @@
 #### [unreleased]
 
+#### 2.1.0 / 2020-07-20
+* update block build process
+* add title and description attributes for a11y
+* add PDF svg for block
+
 #### 2.0.5 / 2020-03-09
 * minor updates to plugin structure on GitHub
 * update tested to